Tests/LibWeb/Text/expected/HTML/HTMLTableCellElement-cellIndex-attribute.txt

lone td.cellIndex = -1
parented td.cellIndex = 0
lone th.cellIndex = -1
parented th.cellIndex = 0
